First pass at MODCXEKB-4
Created an RM API configuration class that can be invoked statically to
return an RMAPIConfiguration object that contains the customer ID and
API key for the tenant making the API call.

Also put the invocation of the config class in each codex-instances API
as an example of how to use it.