Follow on to MODCXEKB-4
Figured out how to test the RMAPIConfiguration class via JUnit/Vertx
unit tests using mocks. It is a little hacky, but it does work. Regular
tests with Rest-Assured via the APIs should be a lot less hacky or not
hacky at all, but since I needed to execute the mocks via a private
HttpClient instance I had to hack it a bit. So these tests are fragile
until I can discover if there is a better way to do what I want to do in
this test.