Towards MODCXEKB-29
Align with mod-codex-mock. Use the came descriptor ID and the
"interfaceType" of "multiple" to allow for module ID based usage of the
interface. This will probably evolve as we integrate.

Without the "multiple" field, the "X-Okapi-Module-Id" will prevent okapi
from finding the interface.

The RMAPIConfiguration class needs to remove the header before calling
mod-configuration or else okapi cannot route the request.

The downside is that the caller must supply the X-Okapi-Module-Id header
set to the module ID of mod-codex-ekb in order for okapi to route it to
mod-codex-ekb. If it is not present you'll get: "No suitable module
found for path /codex-instances"