Implement MODCXEKB-38
We now will perform as many calls to the RM API as the CQL2RMAPI query
parser returns. Each call will be executed and combined in order of
execution when all are complete. The list is then trimmed of any
unrequested entries (due to codex to RM API paging discrepancies).

Also removed *Async calls from CompletableFuture operations. These calls
are performed on threads from the fork/join pool and not Vertx owned.
This can cause problems, so we will continue executing on threads that
are owned by vertx.