PACT: MODCXEKB-8
With feedback from @magdazacharska and better knowledge of how PACT is
going to be used, restructured the PACT test to do the following:
* Use RMAPIConfiguration as the PACT consumer. Previously, the consumer
was RestAssured. Now if the RMAPIConfiguration is updated with contract
breaking changes, this test should fail.
* Generate the JSON body for the contract/return response of the Mock
PACT provider. Previously, the mock data was a static file. This is
fine, however it creates a very strict contract that the provider may
have difficulty honoring for PACT verification. By using the PACT DSL to
generate the body, we can relax certain properties that we don't care
what the values are (such as the metadata).
* Run the test within the Vertx context so that RMAPIConfiguration can
do its thing. Fortunately, we don't need to run the vertical for the
test to pass. RMAPIConfiguration uses the PACT Mock provider's URL as
the Okapi server.