PACT: MODCXEKB-8
Running tests in isolation is a problem. :) When this test is executed
as part of the whole JUnit suite, it fails due to a persistent state
change in the HttpClientFactory. Once the factory is loaded it sets
itself to either mock or regular mode. This is a private static state
and cannot be changed through normal means.

To work around this, we now use reflection and change the value so we
can get a non-mocked HttpClientFactory for RMAPIConfiguration. When the
test completes, we set it back to whatever it was originally.