Fix tests that could have failures due to a bind exception
Some tests use random ephemeral ports in the 49152-65535 range. The odds
of a collision are low, but I have seen it happen occasionally in my dev
environment. While, we can't eliminate the problem given how Vert.x is
started (AFAICT), we can reduce the probability of using a bound port
significantly by first trying to bind to the port. To this end, a
utility class has been added that has a static method to retrieve random
ports that have this high probability of not being in use. Tests that
require random ports should use this utility method to acquire ports.