Implement MODCXEKB-52 - Search by ID
Added support to search by ID when the CQL parser indicates that the
search field is "id". In this case, since there is no mechanism to
search by KBID in the RM API and it is not possible or practical to
retrieve all titles and check the KBID sequentially, we do not make the
RM API call to search titles, rather we make the RM API call to retrieve
a title by ID.

If the title is found, we convert to an instance and store the instance
in an InstanceCollection of length 1. If the title is not found, we
return an InstanceCollection of length 0.

Also fixed a problem with the PACT test affecting the FOLIO
HttpClientFactory mocking setting depending on the order the tests are
executed. We now assume mocking is enabled whenever the tests are run
(the PACT test needs to disable mocking to execute properly, however
when complete it will now always turn it back on, even if it was not on
prior to the test, i.e. the PACT test was the first test executed).