Fix MODCXEKB-56 and default "selection=selected" for RMAPI title search
We now ignore ext context set fields that are unknown to mod-codex-ekb.
Also, when ext.selected is not passed in the CQL, we send the query
arg selection=selected to RM API. This was a change requested by the
SIG. By default, if selection is not passed to the RM API,
"selection=all" is assumed. This is not the behavior SIG members
expected.

Unit tests have also been updated to reflect these changes.