Fixes for MODCXEKB-61 and MODCXEKB-63
With the recent update to the ramls submodule, the API "limit" query argument was increased from 100 to Integer.MAX_VALUE. When paging RM API requests, we relied on the API limit of 100 to limit RM API queries as well. Since the limit can be any positive integer, we need to change RM API paging in order to issue as many RM API queries as needed to fulfill the codex request.

Since mod-codex-ekb depends on mod-configuration, the ModuleDescriptor has been updated to include the "requires" field explicitly setting this dependency. Also added module permissions for implicit configuration permissions required to execute the API.

This commit also marks release 1.0.0.